+++ pkg/artifactory/provider/provider.go
@@ -84,6 +84,7 @@ func Provider() *schema.Provider {
 		"artifactory_ldap_group_setting":                  configuration.ResourceArtifactoryLdapGroupSetting(),
 		"artifactory_backup":                              configuration.ResourceArtifactoryBackup(),
 		"artifactory_repository_layout":                   configuration.ResourceArtifactoryRepositoryLayout(),
+		"artifactory_property_set":                        configuration.ResourceArtifactoryPropertySet(),
 	}
 
 	for _, repoType := range local.RepoTypesLikeGeneric {