+++ scripts/get-access-key.sh
@@ -18,10 +18,19 @@ function getAccessKey() {
   access_token=$(echo "${cookies}" | grep ACCESSTOKEN | awk '{print $7 }')
 
   local access_key
+  local scoped_access_key
   access_key=$(curl -s -g --request GET "${url}/ui/api/v1/system/security/token?services[]=all" \
                       --header "accept: application/json, text/plain, */*" \
                       --header "x-requested-with: XMLHttpRequest" \
                       --header "cookie: ACCESSTOKEN=${access_token}; REFRESHTOKEN=${refresh_token}")
 
-  echo "${access_key}"
+  scoped_access_key=$(curl --location --request POST "${url}/access/api/v1/tokens" \
+                      --header "Authorization: Bearer ${access_key}" \
+                      --header "Content-Type: application/x-www-form-urlencoded" \
+                      --data-urlencode "expires_in=3600" \
+                      --data-urlencode "username=admin" \
+                      --data-urlencode "scope=applied-permissions/user" \
+                      --data-urlencode "description=Created_with_script_in_TF_provider" | jq .access_token)
+
+  echo "${scoped_access_key}"
 }